requirements.txt
py-clob-client>=0.34.0
metar>=1.11.0
httpx>=0.27.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
python-dateutil>=2.9.0
pytz>=2024.1
structlog>=24.0.0
web3>=6.0.0
